refactor(ContainerInstance): remove reflection
BREAKING CHANGE: Remove runtime reflection from ContainerInstance.

This is part of a broader effort to remove reflection
from the library altogether.
This is done as TypeScript 5 does not support
parameter decorators and, as such, code for
managing parameter decorators is pointless.
Therefore, it is removed.

This, of course, does not currently work as a package.
Inject is still present (however it will be removed) and
replaced with a better-suited construct.