feat(Service): Revamp Services decorator
BREAKING CHANGE: Change the way services are declared through
the Service decorator.  The new API takes in an object containing
metadata, with no use of reflection or handlers.

As can be seen in the commit, the Service decorator has 3 different
ways to call it and provide dependencies to it.  A shorthand form
is introduced which allows for terse declaration of individual services.

Further, if a Service with the given ID has already been declared, the
decorator throws an error.  This prevents two services from being
conjoined together, which is a source of confusing behaviour.

Old options such as "referencedBy" are removed.
Moreover, the type wrappers for individual dependencies are
checked in the decorator to ensure they resolve correctly.